f: channel uint in tests

+++ tests_manual/system/manual_iface_tests.cpp
@@ -92,7 +92,7 @@ TEST_CASE("start ap test"){
         RunStatus rs;
         auto base_actor = ActorPtr(make_shared<Actor_Config_sim>());
         base_actor->set(SK::iface, base_iface);
-        REQUIRE_NOTHROW(start_ap(rs, ap_iface, base_actor, {TestConfig::channel, WifiBand::BAND_2_4_or_5, nullopt}, beacon, TestConfig::mac_addr));
+        REQUIRE_NOTHROW(start_ap(rs, ap_iface, base_actor, {static_cast<uint8_t>(TestConfig::channel), WifiBand::BAND_2_4_or_5, nullopt}, beacon, TestConfig::mac_addr));
 
         stop_ap(ap_iface, nullopt);
         log(LogLevel::INFO, "AP stopped");
@@ -120,7 +120,7 @@ TEST_CASE("STA connected to AP in different namespaces") {
         ap_actor->set(SK::iface, ap_phys_iface);
         ap_actor->set(SK::netns, ap_ns);
 
-        start_ap(rs, ap_vif, ap_actor, {TestConfig::channel, WifiBand::BAND_2_4_or_5, nullopt} , beacon,  TestConfig::mac_addr);
+        start_ap(rs, ap_vif, ap_actor, {static_cast<uint8_t>(TestConfig::channel), WifiBand::BAND_2_4_or_5, nullopt} , beacon,  TestConfig::mac_addr);
         log(LogLevel::INFO, "AP started in namespace: {}", ap_ns);
 
         stop_ap(ap_vif, ap_ns);
